added permission check for all resources

+++ resources/action.py
@@ -5,6 +5,7 @@ from models import ActionModel
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
 from flask_jwt_extended import jwt_required
+from resources.permissions import PermissionValidate
 
 blp = Blueprint("Actions", __name__, description="Operations on actions")
 
@@ -15,11 +16,17 @@ class Action(MethodView):
     @jwt_required()
     @blp.response(200, ActionSchema)
     def get(cls, action_id):
+        if not PermissionValidate().get("/action/<string:action_id>", "GET"):
+            abort(401,message=f'Authorization rejected for /action/<string:action_id>, action GET')
+
         action = ActionModel.query.get_or_404(action_id)
         return action
 
     @jwt_required()
     def delete(cls, action_id):
+        if not PermissionValidate().get("/action/<string:action_id>", "DELETE"):
+            abort(401,message=f'Authorization rejected for /action/<string:action_id>, action DELETE')
+
         action = ActionModel.query.get_or_404(action_id)
         db.session.delete(action)
         db.session.commit()
@@ -29,6 +36,9 @@ class Action(MethodView):
     @blp.arguments(ActionUpdateSchema)
     @blp.response(200, ActionSchema)
     def put(self, action_data, action_id):
+        if not PermissionValidate().get("/action/<string:action_id>", "PUT"):
+            abort(401,message=f'Authorization rejected for /action/<string:action_id>, action PUT')
+
         action = ActionModel.query.get(action_id)
 
         if action:
@@ -48,12 +58,18 @@ class ActionList(MethodView):
     @jwt_required()
     @blp.response(200, ActionSchema(many=True))
     def get(cls):
+        if not PermissionValidate().get("/action", "GET"):
+            abort(401,message=f'Authorization rejected for /action, action GET')
+
         return ActionModel.query.all()
 
     @jwt_required()
     @blp.arguments(ActionSchema)
     @blp.response(201, ActionSchema)
     def post(cls, action_data):
+        if not PermissionValidate().get("/action", "POST"):
+            abort(401,message=f'Authorization rejected for /action, action POST')
+
         action=ActionModel(**action_data)
         try:   
             db.session.add(action)
