+++ resources/group.py
@@ -5,10 +5,9 @@ from models import GroupModel,UserModel, PermissionModel
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
 from flask_jwt_extended import jwt_required
+from resources.permissions import PermissionValidate
 
 blp = Blueprint("Groups", __name__, description="Operations on groups")
-
-
 
 
 @blp.route("/group/<string:group_id>")
@@ -16,11 +15,17 @@ class Group(MethodView):
     @jwt_required()
     @blp.response(200, GroupSchema)
     def get(cls, group_id):
+        if not PermissionValidate().get("/group/<string:group_id>", "GET"):
+            abort(401,message=f'Authorization rejected for /group/<string:group_id>, action GET')
+
         group = GroupModel.query.get_or_404(group_id)
         return group
 
     @jwt_required()
     def delete(cls, group_id):
+        if not PermissionValidate().get("/group/<string:group_id>", "DELETE"):
+            abort(401,message=f'Authorization rejected for /group/<string:group_id>, action DELETE')
+
         group = GroupModel.query.get_or_404(group_id)
         db.session.delete(group)
         db.session.commit()
@@ -30,6 +35,9 @@ class Group(MethodView):
     @blp.arguments(GroupUpdateSchema)
     @blp.response(200, GroupSchema)
     def put(self, group_data, group_id):
+        if not PermissionValidate().get("/group/<string:group_id>", "PUT"):
+            abort(401,message=f'Authorization rejected for /group/<string:group_id>, action PUT')
+
         group = GroupModel.query.get(group_id)
 
         if group:
@@ -48,12 +56,18 @@ class GroupList(MethodView):
     @jwt_required()
     @blp.response(200, GroupSchema(many=True))
     def get(cls):
+        if not PermissionValidate().get("/group", "GET"):
+            abort(401,message=f'Authorization rejected for /group, action GET')
+
         return GroupModel.query.all()
 
     @jwt_required()
     @blp.arguments(GroupSchema)
     @blp.response(201, GroupSchema)
     def post(cls, group_data):
+        if not PermissionValidate().get("/group", "POST"):
+            abort(401,message=f'Authorization rejected for /group, action POST')
+
         group=GroupModel(**group_data)
         try:   
             db.session.add(group)
@@ -67,11 +81,13 @@ class GroupList(MethodView):
 @blp.route("/group/<int:group_id>/user/<int:user_id>")
 class LinkUsersToGroup(MethodView):
     @jwt_required()
-    @blp.response(201, UserSchema)
+    @blp.response(201, GroupSchema)
     def post(self, group_id, user_id ):
+        if not PermissionValidate().get("/group/<int:group_id>/user/<int:user_id>", "POST"):
+            abort(401,message=f'Authorization rejected for /group/<int:group_id>/user/<int:user_id>, action POST')
+
         group = GroupModel.query.get_or_404(group_id)
         user = UserModel.query.get_or_404(user_id)
-
         group.users.append(user)
 
         try:
@@ -83,8 +99,11 @@ class LinkUsersToGroup(MethodView):
         return group
 
     @jwt_required()
-    @blp.response(200, UsersAndGroupsSchema)
-    def delete(self, group_id, user_id):    
+    @blp.response(200, GroupSchema)
+    def delete(self, group_id, user_id): 
+        if not PermissionValidate().get("/group/<int:group_id>/user/<int:user_id>", "DELETE"):
+            abort(401,message=f'Authorization rejected for /group/<int:group_id>/user/<int:user_id>, action DELETE')
+
         group = GroupModel.query.get_or_404(group_id)
         user = UserModel.query.get_or_404(user_id)
         group.users.remove(user)
@@ -108,9 +127,11 @@ class LinkGroupsToUser(MethodView):
     @jwt_required()
     @blp.response(201, GroupSchema)
     def post(self, group_id, permission_id):
+        if not PermissionValidate().get("/group/<int:group_id>/permission/<int:permission_id>", "POST"):
+            abort(401,message=f'Authorization rejected for /group/<int:group_id>/permission/<int:permission_id>, action POST')
+
         group = GroupModel.query.get_or_404(group_id)
         permission = PermissionModel.query.get_or_404(permission_id)
-
         group.permissions.append(permission)
 
         try:
@@ -124,9 +145,11 @@ class LinkGroupsToUser(MethodView):
     @jwt_required()
     @blp.response(200, GroupSchema)
     def delete(self, group_id, permission_id):
+        if not PermissionValidate().get("/group/<int:group_id>/permission/<int:permission_id>", "DELETE"):
+            abort(401,message=f'Authorization rejected for /group/<int:group_id>/permission/<int:permission_id>, action DELETE')
+
         group = GroupModel.query.get_or_404(group_id)
         permission = PermissionModel.query.get_or_404(permission_id)
-
         group.permissions.remove(permission)
 
         try:
